refactor(repo): moved mobile and api into packages
