--- mobile/metro.config.js
@@ -1,9 +1,0 @@
-const { getDefaultConfig } = require("expo/metro-config");
-
-const config = getDefaultConfig(__dirname);
-
-config.resolver.alias = {
-	"@": __dirname,
-};
-
-module.exports = config;
