--- mobile/tsconfig.json
@@ -1,13 +1,0 @@
-{
-	"extends": "../tsconfig.base.json",
-	"compilerOptions": {
-		"strict": true,
-		"module": "nodenext",
-		"baseUrl": ".",
-		"jsx": "react-jsx",
-		"paths": {
-			"@/*": ["./*"]
-		}
-	},
-	"include": ["**/*.ts", "**/*.tsx", ".expo/types/**/*.ts", "expo-env.d.ts"]
-}
