--- packages/app/biome.json
@@ -1,3 +1,0 @@
-{
-	"extends": "//"
-}
