--- packages/app/package.json
@@ -1,12 +1,0 @@
-{
-	"name": "@safe-step/app",
-	"version": "0.0.1",
-	"description": "",
-	"main": "index.js",
-	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "ISC"
-}
