+++ packages/mobile/app/(tabs)/explore.tsx
@@ -1,11 +1,11 @@
 import { Image } from "expo-image";
 import { Platform, StyleSheet } from "react-native";
 
+import { Collapsible } from "@/components/ui/collapsible";
 import { ExternalLink } from "@/components/external-link";
 import ParallaxScrollView from "@/components/parallax-scroll-view";
 import { ThemedText } from "@/components/themed-text";
 import { ThemedView } from "@/components/themed-view";
-import { Collapsible } from "@/components/ui/collapsible";
 import { IconSymbol } from "@/components/ui/icon-symbol";
 import { Fonts } from "@/constants/theme";
