+++ packages/mobile/app/(tabs)/index.tsx
@@ -1,11 +1,11 @@
 import { Image } from "expo-image";
-
-import { Link } from "expo-router";
 import { Platform, StyleSheet } from "react-native";
+
 import { HelloWave } from "@/components/hello-wave";
 import ParallaxScrollView from "@/components/parallax-scroll-view";
 import { ThemedText } from "@/components/themed-text";
 import { ThemedView } from "@/components/themed-view";
+import { Link } from "expo-router";
 
 export default function HomeScreen() {
 	return (
