+++ packages/mobile/assets/images/android-icon-background.png
