+++ packages/mobile/assets/images/android-icon-foreground.png
