+++ packages/mobile/assets/images/android-icon-monochrome.png
