+++ packages/mobile/assets/images/favicon.png
