+++ packages/mobile/assets/images/icon.png
