+++ packages/mobile/assets/images/react-logo.png
