+++ packages/mobile/assets/images/react-logo@2x.png
