+++ packages/mobile/assets/images/splash-icon.png
