+++ packages/mobile/components/external-link.tsx
@@ -1,9 +1,9 @@
-import { Link, type Href } from "expo-router";
+import { Href, Link } from "expo-router";
 import {
 	openBrowserAsync,
 	WebBrowserPresentationStyle,
 } from "expo-web-browser";
-import type { ComponentProps } from "react";
+import { type ComponentProps } from "react";
 
 type Props = Omit<ComponentProps<typeof Link>, "href"> & {
 	href: Href & string;
