+++ packages/mobile/components/haptic-tab.tsx
@@ -1,13 +1,12 @@
-import type { BottomTabBarButtonProps } from "@react-navigation/bottom-tabs";
+import { BottomTabBarButtonProps } from "@react-navigation/bottom-tabs";
 import { PlatformPressable } from "@react-navigation/elements";
-import type { GestureResponderEvent } from "react-native";
 import * as Haptics from "expo-haptics";
 
 export function HapticTab(props: BottomTabBarButtonProps) {
 	return (
 		<PlatformPressable
 			{...props}
-			onPressIn={(ev: GestureResponderEvent) => {
+			onPressIn={(ev) => {
 				if (process.env.EXPO_OS === "ios") {
 					// Add a soft haptic feedback when pressing down on the tabs.
 					Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
