+++ packages/mobile/components/ui/icon-symbol.ios.tsx
@@ -1,9 +1,5 @@
-import {
-	SymbolView,
-	type SymbolViewProps,
-	type SymbolWeight,
-} from "expo-symbols";
-import type { StyleProp, ViewStyle } from "react-native";
+import { SymbolView, SymbolViewProps, SymbolWeight } from "expo-symbols";
+import { StyleProp, ViewStyle } from "react-native";
 
 export function IconSymbol({
 	name,
