+++ packages/mobile/hooks/use-theme-color.ts
