+++ pnpm-workspace.yaml
@@ -1,13 +1,6 @@
 packages:
   - packages/*
-  - api
-  - mobile
 
 catalog:
-  '@biomejs/biome': 2.3.4
+  "@biomejs/biome": 2.3.4
   typescript: ^5.9.3
-
-nodeLinker: hoisted
-
-onlyBuiltDependencies:
-  - supabase
